implement createFullBinaryTree
This is an auxiliary function that allows the creation of a
binary search tree for a specified number of levels and with the
end level filled with each individual character of the string
passed as the second argument. It is useful to create binary
search trees for the base 64 encoding so that the correct
character to represent a 6-bit stream can be found in log(n)
time.